requirements.txt
# Core Bot Framework
pyrogram>=2.0.106
tgcrypto>=1.2.5
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# HTTP and Async
httpx>=0.25.0
aiohttp>=3.8.0
aiofiles>=23.0.0

# Media Downloading
yt-dlp>=2023.9.24
instaloader>=4.10.0

# Utilities
pydantic>=2.0.0
pyyaml>=6.0.1
brotli>=1.1.0
pycryptodome>=3.19.0

# Development
black>=23.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0